% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyse_backward}
\alias{analyse_backward}
\title{Backward stepwise selection using AIC}
\usage{
analyse_backward(
  x,
  response = NULL,
  vars = NULL,
  family = stats::gaussian(),
  k = 2
)
}
\arguments{
\item{x}{An output of the \code{bootstrap_data} function.}

\item{response}{a character argument denoting the outcome.}

\item{vars}{a character vector denoting the covariates or features to be
screened.}

\item{family}{the error distribution and link function passed to the \code{glm()}
call inside the \code{analyse_backward} function. Default is \code{gaussian}.}

\item{k}{as per the \code{stats::step} definition: "the multiple of the number of
degrees of freedom used for the penalty".}
}
\value{
A nested \code{tibble}.
}
\description{
An application of backward stepwise selection using AIC as the
criterion for stepping down (i.e., from a full model) through the
candidate set of variables.
}
\details{
The present function is a simple wrapper for the \code{stats::step()}
function. There are various commentaries that have been published on the
use of stepwise selection, with a range of interpretations of its utility,
spanning from more generous proponents focussing on the simplicity with
which a reduced model with "insignificant" variables removed, through to
those that highlight how it invalidates most of any statistical inference
that can be drawn from the final fit. Irrespective of the stance taken,
this method is still common practice in many clinical prediction modelling
publications and is thus worth including for illustrating selection
instability.
}
\examples{
data(iswr_stroke)
iswr_stroke \%>\%
  bootstrap_data(10, seed = 1234) \%>\%
  analyse_backward(response = "dead12",
                   vars = c("Gender", "Age", "Diagnosis", "Coma",
                            "Diabetes", "MI", "Hypertension"),
                   family = "binomial")
}
\references{
Harrell, F. E. (2015). Regression Modeling Strategies: With Applications to
Linear Models, Logistic and Ordinal Regression, and Survival Analysis.
Springer International Publishing.
https://books.google.co.uk/books?id=sQ90rgEACAAJ
}
