% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyse_grlasso}
\alias{analyse_grlasso}
\title{Grouped LASSO penalisation for shrinkage with selection}
\usage{
analyse_grlasso(
  x,
  response = NULL,
  vars = NULL,
  family = "gaussian",
  grpreg_list = list(NULL)
)
}
\arguments{
\item{x}{An output of the \code{bootstrap_data} function.}

\item{response}{a character argument denoting the outcome.}

\item{vars}{a character vector denoting the covariates or features to be
screened.}

\item{family}{the error distribution and link function passed to the
\code{grpreg()} call inside the \code{analyse_grlasso} function. Currently only
\code{gaussian}, \code{binomial} and \code{poisson} are supported in the \code{grpreg} package
with the present default set as \code{"gaussian"}.}

\item{grpreg_list}{a \code{list} passed to the \code{grpreg::grpreg} call. See
\code{?grpreg::grpreg} for more details.}
}
\description{
An application of Grouped LASSO penalisation. Whilst group LASSO
may be used more generally to pre-specify arbitrary groupings of
variables, it is used here to ensure selection of entire categorical
variables with more than one dummy level.
}
\details{
The present function is a simple wrapper for the \code{grpreg::grpreg()}
function. Grouped LASSO extends the standard LASSO penalty through the
introduction of a grouping index. This therefore applies variable
selection at the group level.
}
\examples{
data(iswr_stroke)
iswr_stroke \%>\%
  bootstrap_data(10, seed = 1234) \%>\%
  analyse_grlasso(response = "dead12",
                  vars = c("Gender", "Age", "Diagnosis", "Coma",
                           "Diabetes", "MI", "Hypertension"),
                  family = "binomial",
                  grpreg_list = list(penalty = "grLasso"))
}
