% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyse_univariate}
\alias{analyse_univariate}
\title{Uni-variate variable selection}
\usage{
analyse_univariate(
  x,
  response = NULL,
  vars = NULL,
  level = 0.05,
  family = stats::gaussian(),
  method = "Wald"
)
}
\arguments{
\item{x}{An output of the \code{bootstrap_data} function.}

\item{response}{a character argument denoting the outcome.}

\item{vars}{a character vector denoting the covariates or features to be
screened.}

\item{level}{a numeric scalar denoting a significance threshold to be used
against each uni-variate test performed. Default is \code{0.05}.}

\item{family}{the error distribution and link function passed to the \code{glm()}
call inside the \code{analyse_univariate} function. Default is \code{gaussian}.}

\item{method}{a character value of the chosen test to be used. Currently
only a Wald Chi-squared test is implemented.}
}
\value{
A nested \code{tibble}.
}
\description{
A univariate screening of covariates against some outcome.
}
\details{
A number of \code{analysis} wrappers have been created to illustrate
the challenge of final variable selection. These can be combined in
sequence as desired by the user to consider combinations of selection per
bootstrap replication. Although discouraged (Moons et al., 2012),
univariate selection is not uncommon, particularly as a way to pre-screen
many covariates prior to some multi-variable fit with possible further
feature selection thereafter. The present function is a simple wrapper for
the \code{aod::wald.test()} function, an implementation of the Wald Chi-squared
test under an assumption of asymptotic (multivariate) normality.
}
\examples{
data(iswr_stroke)
iswr_stroke \%>\% 
  bootstrap_data(10, seed = 1234) \%>\%
  analyse_univariate(response = "dead12",
                     vars = c("Gender", "Age", "Diagnosis", "Coma", 
                              "Diabetes", "MI", "Hypertension"),
                     family = "binomial",         
                     level = 0.05)

}
\references{
Moons, K. G. M., Kengne, A. P., Woodward, M., Royston, P., Vergouwe, Y.,
Altman, D. G., & Grobbee, D. E. (2012). Risk prediction models: I.
Development, internal validation, and assessing the incremental value of
a new (bio)marker. Heart, 98(9), 683–690.
https://doi.org/10.1136/heartjnl-2011-301246
}
