% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_data}
\alias{bootstrap_data}
\title{Bootstrap a data set with replacement}
\usage{
bootstrap_data(data, times, include_original = FALSE, seed = NULL, ...)
}
\arguments{
\item{data}{A data frame from which to generate bootstrap samples.}

\item{times}{A numeric input specifying how many bootstrap draws are to be
made on \code{data}}

\item{include_original}{A logical input. If \code{TRUE} the returned output will
contain a row with the original \code{data}. See details. Default is \code{FALSE}.}

\item{seed}{A numeric input specifying the seed to be used. Default is \code{NULL}}

\item{...}{These dots are for future extensions and must be empty}
}
\value{
A nested \code{tibble}
}
\description{
Bootstrap a data set with replacement
}
\examples{

test_dat <- data.frame(a = 1:10, b = rnorm(10))
bootstrap_data(test_dat, times = 5) 

}
