% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iswr_stroke.R
\docType{data}
\name{iswr_stroke}
\alias{iswr_stroke}
\title{Estonian stroke data}
\format{
\subsection{\code{iswr_stroke}}{

A data frame with 814 rows and 8 columns:
\describe{
\item{Gender}{Patient's gender, coded as a factor variable with levels \code{Male} and \code{Female}}
\item{Age}{Patients age recorded as an integer}
\item{Diagnosis}{Patient's diagnosis denoting the type of stroke: \code{ICH}, \code{ID}, \code{INF}, and \code{SAH}}
\item{Coma}{Whether the patient fell into a coma at the time of the original stroke, coded as a factor variable with levels  \code{Yes} or \code{No}}
\item{Diabetes}{Patient's history of diabetes, coded as a factor variable with levels \code{Yes} and \code{No}}
\item{MI}{Patient's history of myocardial infarction, coded as a factor variable with levels \code{Yes} and \code{No}}
\item{Hypertension}{Patient's history of Hypertension, coded as a factor variable with levels \code{Yes} and \code{No}}
\item{dead12}{Denoting whether the patient had died by 12 months or not, coded as a numeric variable \code{1 = yes}, \code{0 = no}}
}
}
}
\source{
\url{https://cran.r-project.org/web/packages/ISwR/index.html}
}
\usage{
iswr_stroke
}
\description{
Taken from the package \code{{ISwR}}, this example dataset is of all cases of
stroke in Tartu, Estonia, during the period 1991–1993, with follow-up until
January 1, 1996. The main edits here are to decorate this with specified
variable types and define a binary classification of "dead or alive" at 12
months from initial stroke.
}
\keyword{datasets}
