% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.tetris_analysis}
\title{Plot of bootstrapped variable selections}
\usage{
plot(x, ...)

\method{plot}{tetris_analysis}(x, k = 0, ...)
}
\arguments{
\item{x}{an output of an \code{analysis_} function of class \code{tetris_analysis}}

\item{k}{a numeric input used to truncate the number of columns in the
"tetris plot". Defaults to \code{0} meaning no restriction}
}
\value{
A \code{patchwork} of two \code{ggplot} outputs
}
\description{
Produces a "tetris plot" highlighting the uncertainty in the
selection of variables chosen under repeat sampling
}
\details{
In some cases there may be a large number of combinations along
the columns of the plot (joint selection probabilities). Such scenarios
can occur when there are many joint selection combinations of empirical
size \code{1/times} where \code{times} are the number of bootstrap draws.  Under
those circumstances it might be desired to restrict to a "top set" which
ignores those less frequent combinations - effectively "zooming" in on the
left-hand-side of the "tetris plot".
}
\examples{

# This example is illustrative only. It should be expected that categorical 
# and binary variables are coded as factors etc as required. 
datasets::mtcars \%>\%
  bootstrap_data(50, seed = 1234) \%>\%
  analyse_univariate(response = "mpg",
                     vars = c("cyl", "disp", "hp", "drat", "wt",
                              "qsec", "vs", "am", "gear", "carb"),
                     level = 0.01) \%>\%
  plot()
}
