% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summary}
\alias{summary}
\alias{summary.tetris_analysis}
\title{Summarise bootstrapped variable selection}
\usage{
summary(x, type = "marginal", ...)

\method{summary}{tetris_analysis}(x, type = "marginal", ...)
}
\arguments{
\item{x}{an output of an \code{analysis_} function of class \code{tetris_analysis}}

\item{type}{a character input of either \code{"marginal"} (default) or \code{"joint"}}
}
\value{
A \code{tibble} with appended joint or marginal probabilities
}
\description{
Alongside a plot, it can also be useful to extract the numbers
behind the "tetris plot".
}
\details{
The summary method essentially derives two types of  empirical
results associated with the generated bootstrap analyses
}
\examples{

# This example is illustrative only. It should be expected that categorical 
# and binary variables are coded as factors etc as required. 
datasets::mtcars \%>\%
  bootstrap_data(50, seed = 1234) \%>\%
  analyse_univariate(response = "mpg",
                     vars = c("cyl", "disp", "hp", "drat", "wt",
                              "qsec", "vs", "am", "gear", "carb"),
                     level = 0.01) \%>\%
  summary()
}
