% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summary_joint_selection}
\alias{summary_joint_selection}
\title{Summarise bootstrapped data}
\usage{
summary_joint_selection(x)
}
\arguments{
\item{x}{An output of the workflow}
}
\value{
A \code{tibble} with appended joint or marginal probabilities
}
\description{
Summarise bootstrapped data
}
\examples{

datasets::mtcars \%>\% 
  bootstrap_data(10) \%>\% 
  analyse_univariate(response = "mpg",
                     vars = c("disp", "hp","drat", "wt","qsec"),
                     level = 0.05) \%>\% 
 summary_joint_selection()                    

}
