% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analyse_univariate}
\alias{analyse_univariate}
\title{Uni-variate variable selection}
\usage{
analyse_univariate(
  x,
  response = NULL,
  vars = NULL,
  level = 0.05,
  family = stats::gaussian(),
  method = "Wald"
)
}
\arguments{
\item{x}{An output of the \code{bootstrap_data} function.}

\item{response}{a character argument denoting the outcome}

\item{vars}{a character vector denoting the covariates or features to be
screened}

\item{level}{a numeric scalar denoting a significance threshold to be used
against each uni-variate test performed. Default is \code{0.05}.}

\item{family}{the error distribution and link function passed to the \code{glm()}
call inside the \code{analyse_univariate} function.}

\item{method}{a character value of the chosen test to be used. Currently
only a Wald chi-squared test is implemented.}
}
\value{
A nested \code{tibble}
}
\description{
Uni-variate variable selection
}
\details{
A number of \code{analysis_} wrappers have been created to illustrate
the challenge of final variable selection. These can be combined as
desired by the user.
}
\examples{

datasets::mtcars \%>\% 
  bootstrap_data(10) \%>\% 
  analyse_univariate(response = "mpg",
                     vars = c("disp", "hp","drat", "wt","qsec"),
                     level = 0.05)

}
