% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summary}
\alias{summary}
\alias{summary.tetris_analysis}
\title{Summarise bootstrapped variable selection}
\usage{
summary(x, type = "marginal", ...)

\method{summary}{tetris_analysis}(x, type = "marginal", ...)
}
\arguments{
\item{x}{an output of an \code{analysis_} function of class \code{tetris_analysis}}

\item{type}{a character input of either \code{"marginal"} (default) or \code{"joint"}}

\item{...}{Currently unused.}
}
\value{
A \code{tibble} with appended joint or marginal probabilities
}
\description{
Alongside a plot, it can also be useful to extract the numbers
behind the "tetris plot".
}
\details{
The summary method essentially derives two types of  empirical
results associated with the generated bootstrap analyses
}
\examples{
data(iswr_stroke)
iswr_stroke \%>\%
  bootstrap_data(10, seed = 1234) \%>\%
  analyse_univariate(response = "dead12",
                     vars = c("Gender", "Age", "Diagnosis", "Coma",
                              "Diabetes", "MI", "Hypertension"),
                     family = "binomial") \%>\% 
  summary()
}
